move socials constant to component/constants.js
