+++ src/components/About/Header/MobileNav.js
@@ -1,5 +1,6 @@
 import React from 'react'
-import { LINKS, SOCIALS } from '../constants'
+import { LINKS } from '../constants'
+import { SOCIALS } from '../../constants'
 
 const MobileNav = ({ setShowNav }) => {
   return (
