+++ src/components/About/constants.js
@@ -14,22 +14,3 @@ export const PROFESSIONAL_EXPERIENCE = [
     title: 'Cybersecurity Analyst',
   },
 ]
-
-export const SOCIALS = [
-  {
-    icon: 'ri-discord-line',
-    link: 'https://discord.gg/PA5ZxAuSp3',
-  },
-  {
-    icon: 'ri-linkedin-box-line',
-    link: 'https://www.linkedin.com/in/jagannathbhatjb/',
-  },
-  {
-    icon: 'ri-github-line',
-    link: 'https://github.com/jagannathBhat',
-  },
-  {
-    icon: 'ri-gitlab-line',
-    link: 'https://gitlab.com/jagannathBhat',
-  },
-]
