+++ src/components/Footer/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { SOCIALS } from '../About/constants'
+import { SOCIALS } from '../constants'
 
 const Footer = () => (
   <div className='flex items-center justify-between p-8'>
