+++ src/components/constants.js
@@ -1,0 +1,18 @@
+export const SOCIALS = [
+  {
+    icon: 'ri-discord-line',
+    link: 'https://discord.gg/PA5ZxAuSp3',
+  },
+  {
+    icon: 'ri-linkedin-box-line',
+    link: 'https://www.linkedin.com/in/jagannathbhatjb/',
+  },
+  {
+    icon: 'ri-github-line',
+    link: 'https://github.com/jagannathBhat',
+  },
+  {
+    icon: 'ri-gitlab-line',
+    link: 'https://gitlab.com/jagannathBhat',
+  },
+]
